% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCityNames.R
\name{addCityNames}
\alias{addCityNames}
\title{Add City Names
Matches coordinates (poins) of cities to urban center polygons.}
\usage{
addCityNames(poly, cities)
}
\arguments{
\item{poly}{SpatialPolygonsDataFrame object (identifying cities as features)}

\item{cities}{path to shapefile containing official city coordinates}
}
\value{
SpatialPolygonsDataFrame object
}
\description{
Add City Names
Matches coordinates (poins) of cities to urban center polygons.
}
\details{
...
}
\examples{
# first get the cities polygons
PATH <- "_misc/GHS_SMOD_POP2015_GLOBE_R2016A_54009_1k_v1_0/GHS_SMOD_POP2015_GLOBE_R2016A_54009_1k_v1_0.tif"
poly <- getCountryCities(PATH, country="SWITZERLAND", tol=0.05)
# load cities from http://worldmap.harvard.edu/data/geonode:placemarks_edited_columns_yz6
cities <- "_misc/placemarks_edited_columns_yz6/placemarks_edited_columns_yz6.shp"
swiss_cities_named <- addCityNames(poly, cities)
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
