% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addressComponents.R
\name{addressComponents}
\alias{addressComponents}
\title{Extract Address Components}
\usage{
addressComponents(r)
}
\arguments{
\item{r}{the "results" part of a detailsRequest response}
}
\value{
a data frame
}
\description{
Extract all address_components of a details search response
}
\details{
Used as internal function in extractJsonDetails()
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
