% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaScan.R
\name{areaScan}
\alias{areaScan}
\title{Perform Radar Searches Over Large Area}
\usage{
areaScan(area, radius, types, details=FALSE)
}
\arguments{
\item{area}{a 2x2 data frame containing the x-coordinates (long) of two points in the first column and the
y-coordinates (lat) in the second column.}

\item{radius}{numeric, the radius for the google places radar search}

\item{types}{character, the location types to be searched for}

\item{details}{logical, indicating whether details on locations should be extracted}
}
\value{
object of class "locations"
}
\description{
Divides the area into equally sized squares and performs a radar search in the center of each square.
}
\details{
The function divides the area into equally sized squares with a diagonal of 2*radius
and performs a radar search in the center of each square (and additionally
a details search for each radar search reference)
}
\examples{
# example of search: scan all switzerland
cantons <- url("http://gadm.org/data/rda/CHE_adm1.RData") # get a shapefile with swiss admin. borders
print(load(cantons))
# extract map edges
edges <- getEdges(x=gadm)
# plot search area
area <- getPointsArea(edges, radius=20000)
plot(area)
scanresp <- areaScan(area=edges, radius=8000, types="restaurant", details=TRUE )
data <- placesData(scanresp)
data <- data[!duplicated(data$id),]
View(data)
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
