% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctern}
\alias{ctern}
\title{Caspian terns plastic debris in Senegal.}
\format{A data frame with 529 rows and 8 variables:
\describe{
  \item{species}{species name, add more info}
  \item{location}{location, add more info}
  \item{country}{country, add more info}
  \item{latitude}{latitude, add more info}
  \item{longitude}{longitude, add more info}
  \item{year}{year, add more info}
  \item{nest_code}{nest code, add more info}
  \item{debris_presence}{debris presence absence, add more info}
}}
\usage{
ctern
}
\description{
A dataset containing absence and presence observations
of plastic debris for the Caspian terns in the coast
of Senegal.
}
\references{
Tavares et al. Submitted.
}
\keyword{datasets}
