% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailsRequest.R
\name{detailsRequest}
\alias{detailsRequest}
\title{Details Request}
\usage{
detailsRequest(url.list)
}
\arguments{
\item{url.list}{list of urls (character strings)}

\item{usage}{detailsRequest(url.list)}
}
\value{
a list with the parsed json objects
}
\description{
Handles GET requests for the details method of the Google Places API
}
\details{
sends GET requests, extracts and parses response objects, returns a list with the parsed json objects
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
