% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDetailsResults.R
\name{extractDetailsResults}
\alias{extractDetailsResults}
\title{Extract Details Results}
\usage{
extractDetailsResults(x)
}
\arguments{
\item{x}{a nested list containing search results from a details search}
}
\value{
a data frame
}
\description{
Extract all results from a detail response.list (all in one data frame)
}
\details{
x: a nested list containing search results from a details search. each main element is a parsed json response from google places api
see extractJsonDetails for what key:value pairs of the parsed json responses are extracted.
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
