% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractJson.R
\name{extractJson}
\alias{extractJson}
\title{Extract Data From Response Object}
\usage{
extractJson(response)
}
\arguments{
\item{response}{nested list, parsed json-object returned from the API}
}
\value{
a data.frame/data.table (containing the extracted data, or the error message from the response)
}
\description{
Internal function to extract the data from one API response
}
\details{
Takes parameter values according to google Places API
all parameter values are character strings, but the location value can be a
list of lat/long-character-strings (allowing for automated search of many locations)
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
