% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractJsonDetails.R
\name{extractJsonDetails}
\alias{extractJsonDetails}
\title{Extract JSON Details}
\usage{
extractJsonDetails(response)
}
\arguments{
\item{response}{the response object of a detailsRequest response}
}
\value{
a data frame
}
\description{
Internal function to extract the data from one details search response
}
\details{
...
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
