% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPolygonIslands.R
\name{extractPolygonIslands}
\alias{extractPolygonIslands}
\title{Extract Polygon Islands}
\usage{
extractPolygonIslands(spdf)
}
\arguments{
\item{spdf}{SpatialPolygonsDataFrame object}
}
\value{
SpatialPolygons object
}
\description{
Store each polygon (island) as individual feature of a SpatialPolygons object
}
\details{
The input spdf contains several polygons, all in one feature (as a result of extracting polygons from a raster).
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
