% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractResultsRadar.R
\name{extractResultsRadar}
\alias{extractResultsRadar}
\title{Extract Data From Radar Search Results}
\usage{
extractResultsRadar(x, l)
}
\arguments{
\item{x}{a nested list containing search results. each main element is a parsed json response from google places api}

\item{l}{the location-input-list for the main search function}
}
\value{
a data.frame/data.table (containing the extracted data from all objects in the response)
}
\description{
Extract all results from a response.list (all in one data frame), adds the location-input (l) to the respective response
}
\details{
see extractJsonRadar for details on what key:value pairs of the parsed json responses are extracted.
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
