% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountryCities.R
\name{getCountryCities}
\alias{getCountryCities}
\title{Get City Polygons of a Country
Extracts polygon 'city' boundaries from GHS settlement grid raster data}
\usage{
getCountryCities(r, country, min_value, tol=0.1)
}
\arguments{
\item{r}{RasterLayer object of imported GHS settlement grid raster file, or path to such a file}

\item{country}{character, name of the country according to GADM database}

\item{min_value}{integer, minimum cell value to consider (as 'city', defaults to 3, urban centers)}

\item{tol}{numeric, tolerance value to be used by the Douglas-Peuker algorithm to simplify the country-shapefile, defaults to .1}
}
\value{
SpatialPolygons object
}
\description{
Get City Polygons of a Country
Extracts polygon 'city' boundaries from GHS settlement grid raster data
}
\details{
The European Commission's GHS SETTLEMENT GRID (LDS) (https://ghslsys.jrc.ec.europa.eu/ghs_smod.php) provides raster data on settlements (urban centers/surroundings)
worldwide. This function takes their data as input, identifies all urban areas ('cities') in a given country, and extracts
the shape of the cities (polygons) as features of one SpatialPolygons object of the country's extent.
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
