% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEdges.R
\name{getEdges}
\alias{getEdges}
\title{Get Polygon Edges}
\usage{
getEdges(x, zoomfactor=1)
}
\arguments{
\item{x}{object of class Polygon, Polygons, SpatialLinesDataFrame, SpatialPolygons, or SpatialPolygonsDataFrame}

\item{zoomfactor}{numeric, a number to multiply the resulting long/lat values in order to "zoom in" or "zoom out" (e.g. a zoomfactor biggar than 1 is widening the area, defaults to 1)}
}
\value{
the function returns an object of class placesearch, containing the search results and information on the search parameters
}
\description{
Returns a 2x2 data frame containing the two extreme points (edges) of the polygon.
}
\details{
Given a map object of class Polygon, Polygons, SpatialLinesDataFrame, SpatialPolygons, or SpatialPolygonsDataFrame
the function returns a 2x2 data frame containing the two extreme points (edges) of the polygon with the x-coordinates (long)  in the first column and the
y-coordinates (lat) in the second column. This is used as input for a area search. (instead of manually providing such a data.frame)
}
\examples{
# example:
cantons <- url("http://gadm.org/data/rda/CHE_adm1.RData")
print(load(cantons))
ag <- gadm[gadm$NAME_1=="Aargau",]
edges <- getEdges(x=ag, zoomfactor=1.01)
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
