% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPointsArea.R
\name{getPointsArea}
\alias{getPointsArea}
\title{Get Cell Centers of large Area}
\usage{
getPointsArea(area, radius)
}
\arguments{
\item{area}{a data.frame containing two lng/lat points}

\item{radius}{numeric, the radius of the individual search areas}
}
\value{
object of class "locations"
}
\description{
Divides an area into cells of the size with the diagonal
and returns the lng/lat coordinates of the cells' center
}
\details{
A function that divides an area (defined by two lng/lat points) into cells of the size with the diagonal
twice the given radius (in meters)  and returns the lng/lat coordinates of the cells' center
area: a 2x2 data frame containing the x-coordinates (long) of two points in the first column and the
y-coordinates (lat) in the second column.
The function returns an object of class "locations" (see classes_methods.R) containing
the lng/lat coordinates as well as additional data for print- and plot methods for location objects.
}
\examples{
# example:
r2squareSide(5)
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
