% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlongString.R
\name{latlongString}
\alias{latlongString}
\title{Get Coordinate String}
\usage{
latlongString(coords)
}
\arguments{
\item{coords}{numeric, coordinates from getPointArea}
}
\value{
character vector with coordinate strings
}
\description{
Transforms coordinats from getPointsArea into character string for searchRadar
}
\details{
a function that takes an x and y coordinate (numbers from a getPointsArea response (df)) and
returns each row as a character string for the searchRadar tpation parameter.
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
