% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{locations}
\alias{locations}
\title{Get Locations Vector}
\usage{
locations(x)
}
\arguments{
\item{x}{object of class location}
}
\value{
character vector with coordinate strings
}
\description{
Takes an object of class locations as an input and returns
the lng/lat-locations-vector of the object's search slot
}
\details{
a function that takes an object of class locations as an input and returns
the lng/lat-locations-vector of the object's search slot.
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
