% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openingHours.R
\name{openingHours}
\alias{openingHours}
\title{Extract Opening Hours}
\usage{
openingHours(r)
}
\arguments{
\item{r}{the "results" part of a detailsRequest response}
}
\value{
a data frame
}
\description{
Extract opening hours of a details search response
}
\details{
a function to extract all opening hours of a response and to transform them in a df with days as columns
containing times as values. Used as internal function in extractJsonDetails()
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
