% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placer.R
\docType{package}
\name{placer}
\alias{placer}
\alias{placer-package}
\title{placer: A package to estimate the accumulation of plastic debris in bird's nests}
\description{
The package placer consist of two main functions to estimate the accumulation of plastic
in bird's nest as a function of sample size, and a routine to plot the plastic prevalence
probability, and their confidence intervals as a function of sample size.
}
\section{Main functions}{

The function \code{\link{plastic.prev.prob}} calculates the plastic
prevalence probability for a given sample size based on presence and absence data
The function \code{\link{plastic.ci}} estimates the 95\% confidence intervals for a given
prevalence probability of plastic debris. In addition, the package placer includes a plotting
routine \code{\link{prevalence_plot}} to show the estimated plastic prevalence probability and
their 95\% confidence intervals as a function of sample size.
}

