% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placesData.R
\name{placesData}
\alias{placesData}
\title{Get Places Data}
\usage{
placesData(x)
}
\arguments{
\item{x}{object of class location}
}
\value{
character vector with coordinate strings
}
\description{
Extracts the search.results from a scan object
}
\details{
A function that extracts the search.results from a scan object.
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
