% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.locations.R
\name{plot.locations}
\alias{plot.locations}
\title{Plot Method For Location Objects}
\usage{
plot.locations(x, ggmap.zoom = 7, ncircles = 1, googlemap = TRUE)
}
\arguments{
\item{x}{object of class location}

\item{ggmap.zoom}{integer, zoom parameter for ggmap output}

\item{ncircles}{integer, the number of circles to be plotted (centered around search area, defaults to 1)}

\item{googlemap}{logical, if TRUE, plots google map as background. Defaults to TRUE.}
}
\value{
NULL (output is plotted)
}
\description{
Takes an object of class locations and plots the locations to a map
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
