% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_methods.R
\name{plot.placesearch}
\alias{plot.placesearch}
\title{Plot a placesearch object on a map}
\usage{
plot.placesearch(
  x,
  ggmap.zoom = "auto",
  ncircles = 1,
  col.grid = "white",
  col.loc = "#e55562",
  col.area = "#6b3a70",
  col.search = "#4db0d9",
  col.circles = "#4db0d9"
)
}
\arguments{
\item{x}{A placesearch object to be plotted on the map.}

\item{ggmap.zoom}{Zoom level for the map (default: "auto").}

\item{ncircles}{Number of circles to be plotted (default: 1).}

\item{col.grid}{Color for the grid (default: "white").}

\item{col.loc}{Color for the location points (default: "#e55562").}

\item{col.area}{Color for the area (default: "#6b3a70").}

\item{col.search}{Color for the search points (default: "#4db0d9").}

\item{col.circles}{Color for the circles (default: "#4db0d9").}
}
\value{
A ggplot object with the plotted placesearch object.
}
\description{
This function takes a placesearch object and plots it on a map using ggmap.
The function allows for customization of colors and zoom level.
}
\examples{
\dontrun{
# Assuming you have a placesearch object named 'ps_obj'
plot.placesearch(ps_obj)
}
}
