% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.scan.R
\name{plot.scan}
\alias{plot.scan}
\title{Plot Method For Scan Objects}
\usage{
plot.scan(x, ggmap.zoom=7, ncircles=1, col.grid="white", col.loc="#e55562", col.area="#6b3a70", col.search="#4db0d9", col.circles="#4db0d9", poly.map = NULL )
}
\arguments{
\item{x}{object of class scan}

\item{ggmap.zoom}{integer, zoom parameter for ggmap output}

\item{ncircles}{integer, the number of circles to be plotted (centered around search area, defaults to 1)}

\item{col.grid}{character, the color of the grid (in hexadecimal or name of R-color; default:"white")}

\item{col.loc}{character, the color of the locations found (in hexadecimal or name of R-color; default: "#e55562")}

\item{col.area}{character, the color of the rectangle indicating the search area (in hexadecimal or name of R-color; default: "#6b3a70")}

\item{col.search}{character, the color of the search area center (in hexadecimal or name of R-color; default: "#4db0d9")}

\item{poly.map}{spatialPolygonDataframe (optional), map data as basis for the plot. If empty (NULL), the function plots a google map as basis.}

\item{col.search}{character, the color of the circles indicating the search radius (in hexadecimal or name of R-color; default: "#4db0d9")}
}
\value{
a ggplot object
}
\description{
Takes an object of class scan and plots the locations and search grid to a map
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
