% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2squareSide.R
\name{r2squareSide}
\alias{r2squareSide}
\title{Get Square Side}
\usage{
r2squareSide(x, zoomfactor=1)
}
\arguments{
\item{r}{numeric, radius of a circle}
}
\value{
numeric
}
\description{
Compute the length of the square side of the square that has a diagonal equal to 2r
}
\details{
Given a radius r of a circle, compute the length of the square
side of the square that has a diagonal equal to 2r
}
\examples{
# example:
r2squareSide(5)
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
