% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveDL.R
\name{saveDL}
\alias{saveDL}
\title{Save Download}
\usage{
saveDL(
  request.function,
  request.id,
  chunksize = 100,
  pause = 0,
  backupfile = "saveDl.list.Rdata"
)
}
\arguments{
\item{request.function}{the request function used for the individual downloads}

\item{request.id}{list, the input to the request function}

\item{chunksize}{ingeger, number of cases in one chunk}

\item{pause}{numeric, duration of break between chunks in seconds}

\item{backupfile}{character, the path/name of the file to cache the data during download in progress}
}
\value{
a list with the parsed json objects
}
\description{
Handles download of API data via large number of requests
}
\details{
an internal function for request functions to manage a large number of requests. the queries are split up and the responses saved on disk chunk by chunk
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
