% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchNearby.R
\name{searchNearby}
\alias{searchNearby}
\title{Search Nearby}
\usage{
searchNearby(location, radius, sensor = "false", keyword = NULL,
  language = NULL, name = NULL, rankby = NULL, types = NULL)
}
\arguments{
\item{location}{character, containing the coordinates (long/lat separated by a comma) of where the map search shall be executed}

\item{radius}{numeric, giving radius to be used for the nearby search}

\item{sensor}{character, the map search sensor option (defaults to "false")}

\item{keyword}{character, keywords to be used (defaults to NULL)}

\item{language}{character, language parameter of search (defaults to NULL, default language will be used)}

\item{name}{character, name parameter of search (defaults to NULL)}

\item{rankby}{character, rank parameter for search results (defaults to NULL)}

\item{type}{character, type of objects to be searched (defaults to NULL)}
}
\value{
the function returns an object of class placesearch, containing the search results and information on the search parameters
}
\description{
Main search function
}
\details{
Takes parameter values according to google Places API
all parameter values are character strings, but the location value can be a
list of lat/long-character-strings (allowing for automated search of many locations)
NOTE the function automatically requests additional results
if available. This requires to suspend execution in order to
respect the time gap "between when a next_page_token is issued, and when it will become valid."
(see: https://developers.google.com/places/documentation/search)
}
\examples{
example (taken from google developers):
searchNearby(location="-33.8670522,151.1957362", rankby="distance", types="food", name="harbour")
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
