% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchRequest.R
\name{searchRequest}
\alias{searchRequest}
\title{Search Request}
\usage{
searchRequest(url.list)
}
\arguments{
\item{url.list}{list of urls (character strings)}
}
\value{
a list with the parsed json objects
}
\description{
sends GET requests, extracts and parses response objects,
}
\details{
NOTE the function automatically requests additional results if available. This requires to suspend execution in order to
respect the time gap "between when a next_page_token is issued, and when it will become valid." (see: https://developers.google.com/places/documentation/search)
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
