% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchRequestNearby.R
\name{searchRequestNearby}
\alias{searchRequestNearby}
\title{Handle Nearby Search}
\usage{
searchRequestNearby(url.list)
}
\arguments{
\item{url.list}{list of urls (character strings)}
}
\value{
a list with the parsed json objects
}
\description{
Sends GET requests, extracts and parses response objects
}
\details{
NOTE the function automatically requests additional results if available. This requires to suspend execution in order to
respect the time gap "between when a next_page_token is issued, and when it will become valid." (see: https://developers.google.com/places/documentation/search)
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
