% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchRequestRadar.R
\name{searchRequestRadar}
\alias{searchRequestRadar}
\title{Search Request Radar}
\usage{
searchRequestRadar(url.list)
}
\value{
a list with the parsed json objects
}
\description{
Sends GET requests, extracts and parses response objects
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
