% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCityNames.R
\name{addCityNames}
\alias{addCityNames}
\title{Add City Names
Matches coordinates (points) of cities to urban center polygons.}
\usage{
addCityNames(poly, cities, expand=100)
}
\arguments{
\item{poly}{SpatialPolygonsDataFrame object (identifying cities as features)}

\item{cities}{path to shapefile containing official city coordinates}

\item{expand}{numeric, the width to slightly expand the polygon for the overlay (helps in coast regions, where the overlay with points might fail). Defaults to 100.}
}
\value{
SpatialPolygonsDataFrame object
}
\description{
Add City Names
Matches coordinates (points) of cities to urban center polygons.
}
\details{
Note that if extend is larger than 0, the extension only affects the overlay with city ponits. The returned polygon is not extended.
}
\examples{
# first get the cities polygons
PATH <- "_misc/GHS_SMOD_POP2015_GLOBE_R2016A_54009_1k_v1_0/GHS_SMOD_POP2015_GLOBE_R2016A_54009_1k_v1_0.tif"
poly <- getCountryCities(PATH, country="ITALY", tol=0.05)
# load cities from http://worldmap.harvard.edu/data/geonode:placemarks_edited_columns_yz6
cities <- "_misc/placemarks_edited_columns_yz6/placemarks_edited_columns_yz6.shp"
italian_cities_named <- addCityNames(poly, cities, country="IT")
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
