% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaScan.R
\name{areaScan}
\alias{areaScan}
\title{Perform Radar Searches Over Large Area}
\usage{
areaScan(area, radius, types, details=FALSE, random=FALSE, random_breaks=NULL)
}
\arguments{
\item{area}{a 2x2 data frame containing the x-coordinates (long) of two points in the first column and the
y-coordinates (lat) in the second column, or a list of such data frames.}

\item{radius}{numeric, the radius for the google places radar search}

\item{types}{character, the location types to be searched for}

\item{details}{logical, indicating whether details on locations should be extracted}

\item{random}{logical, indicating whether the order of points to search should be random (default:FALSE)}

\item{random_breaks}{numeric vector of length 2, indicating the interval [min,max] no. of seconds to wait
between calls to the api, defaults to NULL (no breaks)}
}
\value{
object of class "locations" (if area is one dataframe) or a list of objects of class "locations" (if area is a list of dataframes)
}
\description{
Divides the area into equally sized squares and performs a radar search in the center of each square.
}
\details{
The function divides the area into equally sized squares with a diagonal of 2*radius
and performs a radar search in the center of each square (and additionally
a details search for each radar search reference)
NOTE: randomizations and list-input only works if details==FALSE!
}
\examples{
\dontrun{
# example of search: scan all switzerland
library(sp)
cantons <- readRDS(url("https://biogeo.ucdavis.edu/data/gadm3.6/Rsp/gadm36_CHE_1_sp.rds"))
# extract map edges for two cantons
edges <- getAllEdges(list(cantons[1,], cantons[2,]))
# plot search area
area <- getAllPointsArea(edges, radius=20000)
plot(area[[1]])
plot(area[[2]])
scanresp <- areaScan(area=edges,radius=20000, types="church",details=FALSE, random=TRUE, random_breaks=c(0,0.3))
data <- placesData(scanresp)
data <- data[!duplicated(data$id),]
View(data)
}
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
