% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountryCities.R
\name{getCountryCities}
\alias{getCountryCities}
\title{Get City Polygons of a Country
Extracts polygon 'city' boundaries from GHS settlement grid raster data}
\usage{
getCountryCities(r, country, min_value = 3, tol=0.1, intermediate=FALSE)
}
\arguments{
\item{r}{RasterLayer object of imported GHS settlement grid raster file, or path to such a file}

\item{country}{character, name of the country according to GADM database}

\item{min_value}{integer, minimum cell value to consider (as 'city', defaults to 3, urban centers)}

\item{tol}{numeric, tolerance value to be used by the Douglas-Peuker algorithm to simplify the country-shapefile, defaults to .1}

\item{intermediate}{logical, if TRUE, intermediate results (raster data and country polygon are returned as well), defaults to FALSE}
}
\value{
SpatialPolygonsDataFrame object (or a list of SpatialPolygonsDataFrame, raster and SpatialPolygons objects, see details)
}
\description{
Get City Polygons of a Country
Extracts polygon 'city' boundaries from GHS settlement grid raster data
}
\details{
The European Commission's GHS SETTLEMENT GRID (LDS) (https://ghslsys.jrc.ec.europa.eu/ghs_smod.php) provides raster data on settlements (urban centers/surroundings)
worldwide. This function takes their data as input, identifies all urban areas ('cities') in a given country, and extracts
the shape of the cities (polygons) as features of one SpatialPolygons object of the country's extent.
}
\examples{
PATH <- "_misc/GHS_SMOD_POP2015_GLOBE_R2016A_54009_1k_v1_0/GHS_SMOD_POP2015_GLOBE_R2016A_54009_1k_v1_0.tif"
swiss_cities <- getCountryCities(PATH, country="SWITZERLAND", tol=0.005)
# contains by default the entire city polygons (also over the country border)
plot(swiss_cities)
# only consider parts within country
plot(swiss_cities[swiss_cities$in_country,])
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
