% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEdges.R
\name{getEdges}
\alias{getEdges}
\title{Get Polygon Edges}
\usage{
getEdges(x, zoomfactor=1)
}
\arguments{
\item{x}{object of class Polygon, Polygons, SpatialLinesDataFrame, SpatialPolygons, or SpatialPolygonsDataFrame}

\item{zoomfactor}{numeric, a number to multiply the resulting long/lat values in order to "zoom in" or "zoom out" (e.g. a zoomfactor biggar than 1 is widening the area, defaults to 1)}
}
\value{
the function returns an object of class placesearch, containing the search results and information on the search parameters
}
\description{
Returns a 2x2 data frame containing the two extreme points (edges) of the polygon.
}
\details{
getAllEdges is the vectorized version
Given a map object of class Polygon, Polygons, SpatialLinesDataFrame, SpatialPolygons, or SpatialPolygonsDataFrame
the function returns a 2x2 data frame containing the two extreme points (edges) of the polygon with the x-coordinates (long)  in the first column and the
y-coordinates (lat) in the second column. This is used as input for a area search. (instead of manually providing such a data.frame)
}
\examples{
\dontrun{
# example:
cantons <- readRDS(url("https://biogeo.ucdavis.edu/data/gadm3.6/Rsp/gadm36_CHE_1_sp.rds"))
ag <- cantons[cantons$NAME_1=="Aargau",]
edges <- getEdges(x=ag, zoomfactor=1.01)
several_cantons <- list(cantons[1,], cantons[2,])
several_edges <- getAllEdges(several_cantons)
}
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
