% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeDetails.R
\name{placeDetails}
\alias{placeDetails}
\title{Place Details Function}
\usage{
placeDetails(
  reference,
  language = NULL,
  saveDL = FALSE,
  chunksize = 100,
  pause = 0
)
}
\arguments{
\item{reference}{A character vector of reference(s) for the place(s) you want to retrieve details for.}

\item{language}{(Optional) A character string specifying the language to be used for the returned data. Default is NULL.}

\item{saveDL}{(Optional) A logical value indicating whether to save the responses to disk. Default is FALSE.}

\item{chunksize}{(Optional) An integer specifying the number of requests to process at a time when saving responses to disk. Default is 100.}

\item{pause}{(Optional) A numeric value indicating the time (in seconds) to pause between chunks when saving responses to disk. Default is 0.}
}
\value{
A "placedetails" S4 object containing the search results (results.df) and the parameters used for the request (parameters.df).
}
\description{
This function retrieves detailed information about a place using its reference.
It can also save the results to disk and allows for customizing the request process.
}
\examples{
\dontrun{
reference <- "some_place_reference"
place_details <- placeDetails(reference, language="en", saveDL=FALSE)
}
}
