% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeDetails.R
\name{placeDetails}
\alias{placeDetails}
\title{Get Details Of Place}
\usage{
placeDetails(reference, language=NULL, saveDL=FALSE, chunksize=100, pause=0 )
}
\arguments{
\item{reference}{character, the Google Places API reference code of a place (i.e. a restaurant).}
}
\value{
a placedetails object
}
\description{
Takes the reference code of a place and returns detailed information about this place.
}
\details{
see https://developers.google.com/places/documentation/details for a list of all variables
}
\author{
Ulrich Matter <umatter@protonmail.com>
}
