% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectordestination.R
\name{vectordestination}
\alias{vectordestination}
\title{Draw Circle On Map}
\usage{
vectordestination(lonlatpoint, travelvector)
}
\arguments{
\item{travelvector}{direction of circle}

\item{longlatpoint}{numeric vector with coordinates}
}
\value{
the function returns an object of class placesearch, containing the search results and information on the search parameters
}
\description{
Draw a circle with radius of x meters
}
\details{
# Function written by Arien Lam
}
\examples{
# example: use this function to

radius <- 10000 # in meter
NewHaven <- c(-72.9,41.3) #c(lon,lat)
circlevector <- as.data.frame(cbind(direction = seq(0, 2*pi,
                                                     by=2*pi/100), magnitude = radius))
mycircle <- vectordestination(NewHaven, circlevector)
plot(mycircle,type="l")##' @export

}
\author{
Arien Lam
}
